URIs.jl

URIs is a Julia package for parsing and working with Uniform Resource Identifiers, as defined in RFC 3986.

API Reference

URIs.URI — Type

URI(; scheme="", host="", port="", etc...)
URI(str) = parse(URI, str::String)

A type representing a URI (e.g. a URL). Can be constructed from distinct parts using the various supported keyword arguments, or from a string. The URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

When constructing a URI from a String, you need to first unescape that string: URI( URIs.unescapeuri(str) ).

The URI struct stores the complete URI in the uri::String field and the component parts in the following SubString fields:

scheme, e.g. "http" or "https"

userinfo, e.g. "username:password"

host e.g. "julialang.org"

port e.g. "80" or ""

path e.g "/"

query e.g. "Foo=1&Bar=2"

fragment

The resource(::URI) function returns a target-resource string for the URI RFC7230 5.3. e.g. "$path?$query#$fragment".

The queryparams(::URI) function returns a Dict containing the query.

URIs.resource — Function

"request-target" per https://tools.ietf.org/html/rfc7230#section-5.3

Missing docstring.

Missing docstring for queryparams. Check Documenter's build log for details.

Missing docstring.

Missing docstring for absuri. Check Documenter's build log for details.

URIs.escapeuri — Function

percent-encode a string, dict, or pair for a uri

URIs.unescapeuri — Function

unescape a percent-encoded uri/url

Missing docstring.

Missing docstring for escapepath. Check Documenter's build log for details.

URIs.splitpath — Function

Splits the path into components See: http://tools.ietf.org/html/rfc3986#section-3.3

Base.isvalid — Method

checks if a URI is valid