URIs.jl

URIs is a Julia package for parsing and working with Uniform Resource Identifiers, as defined in RFC 3986.

API Reference

URIs.URI — Type

URI(; scheme="", host="", port="", etc...)
URI(str) = parse(URI, str::String)

A type representing a URI (e.g. a URL). Can be constructed from distinct parts using the various supported keyword arguments, or from a string. The URI constructors will automatically escape any provided query arguments, typically provided as "key"=>"value"::Pair or Dict("key"=>"value"). Note that multiple values for a single query key can provided like Dict("key"=>["value1", "value2"]).

When constructing a URI from a String, you need to first unescape that string: URI( URIs.unescapeuri(str) ).

The URI struct stores the complete URI in the uri::String field and the component parts in the following SubString fields:

scheme, e.g. "http" or "https"

userinfo, e.g. "username:password"

host e.g. "julialang.org"

port e.g. "80" or ""

path e.g "/"

query e.g. "Foo=1&Bar=2"

fragment

The queryparams(::URI) function returns a Dict containing the query.

URIs.queryparams — Function

queryparams(::URI)
queryparams(query_str::AbstractString)

Returns a Dict containing the query parameter string parsed according to the key=value pair formatting convention.

Note that this is not part of the formal URI grammar, merely a common parsing convention — see RFC 3986.

URIs.absuri — Function

absuri(uri, context)

Construct an absolute URI, using uri.path and uri.query and filling in other components from context.

URIs.escapeuri — Function

escapeuri(x)

Apply URI percent-encoding to escape special characters in x.

URIs.unescapeuri — Function

unescapeuri(str)

Percent-decode a string according to the URI escaping rules.

URIs.escapepath — Function

escapepath(path)

Escape the path portion of a URI, given the string path containing embedded / characters which separate the path segments.

URIs.splitpath — Function

Splits the path into components See: http://tools.ietf.org/html/rfc3986#section-3.3

Base.isvalid — Method

checks if a URI is valid